remove logic from Dashboard
